3rd person controller (take 2)
Committing 3rd person controller but without the aspect of trying to
select a minion to take over.  Need to work out what key.  Already have
a similar code to do this from a previous assignment that shouldn't take
too much time to adapt.  Not too sure about having to use the mouse
follow script.  I could make that activate when the Player decides to
take control of a minion and leave it pure 3rd person controls in
between.