Player control mods and box materials applied
Player control has been modified to use the jumping box.  At this stage,
the player must rush the box and NOT jump on it.  It seems very
unreliable when jumping on it but a rush at the box works every time.
The variable than controls his height in the jump is the "BounceHeight"
which can be adjusted in the inspector.

3 materials made and applied to the boxes that Chris made and have been
turned into prefabs.