Half finished Tutorial
Made some things look pretty, made a custom inspector (which is pretty
pointless at the moment until I find a solution to the position) for
TextMesh which adds a big text area that allows you to write larger text
messages easier. Disabled the tutorial at the moment because it's only
half done and don't want it to get in the way.