prefabs and tinkering with line of sight
Just tinkering with the prefabs and mage sure that the ones from the
tute level were put into the prefabs.  tinkering with line of sight.
still sketchy.  fucked if I can work out how these spherecasts work.  It
makes sense when I write the code but the result seems to be "random"