Stopped the minions rotating ..
when you came back to walter.  set the rigidbody freezerotations to true
when they aren't active.  This created a problem with the mini minion
who fell over and hung upside down in the floor.  had to do with his
mesh collider, so I deactivated that ( hope it didn't have another use)